Remove apparently unnecessary code.
In my tests with an Android device attached to a Raspberry pi, SimpleRT
still works with this code removal. However, I am lacking the deeper
knowledge to make assurances that the code is still going to work in all
situations.

I am however pretty sure that the code that I deleted was always pretty
strange and doesn't really make sense to me. The code is calling
libusb_detach_kernel_driver(...). The documentation for
libusb_detach_kernel_driver(...) says that you need to call that function
to be able to claim the interface. However, SimpleRT never did call
libusb_claim_interface(...) _after_ it has detached the kernel driver.
So why would it need to detach it in the first place?

SimpleRT did call libusb_claim_interface(...) in some situations but always
without calling libusb_detach_kernel_driver(...) or
libusb_kernel_driver_active(...) before. So why would it try to claim the
interface without previously making sure that claiming the interface would
actually succeed? My guess is that none of that matters because it isn't
necessary to claim the interface. That's why I deleted those parts of the
code. I would be interested in seeing a situation in which SimpleRT fails
now because it hasn't claimed interface 0. In such a situation the point of
failure would probably be the function call to libusb_control_transfer(...).